ref: Refine API for labeling GPU objects
Previously, the API was inconsistent by trying to juggle `String` and `&'static str`
for labels instead of just using string slices with proper lifetimes.

Furthermore, the labels weren't wrapped in an Option<T>, meaning they had to be provided,
which has been changed in this commit, where they are completely optional now.